requirements.txt
requests
beautifulsoup4
pillow
lxml
toml
